Add x86_64 support

+++ client/cl_scrn.c
@@ -612,11 +612,11 @@ int entitycmpfnc( const entity_t *a, const entity_t *b )
 	*/
 	if ( a->model == b->model )
 	{
-		return ( ( int ) a->skin - ( int ) b->skin );
+		return ( ( uintptr_t ) a->skin - (uintptr_t) b->skin );
 	}
 	else
 	{
-		return ( ( int ) a->model - ( int ) b->model );
+		return ( (uintptr_t) a->model - (uintptr_t) b->model );
 	}
 }
