+++ win32/rw_imp.c
@@ -405,7 +405,7 @@ void SWimp_AppActivate( qboolean active )
 Sys_MakeCodeWriteable
 ================
 */
-void Sys_MakeCodeWriteable (unsigned long startaddr, unsigned long length)
+void Sys_MakeCodeWriteable(uintptr_t startaddr, unsigned long length)
 {
 	DWORD  flOldProtect;
